Vue Tour fixes (#674)
https://github.com/project-chip/zap/pull/625
