+++ src/App.vue
@@ -26,42 +26,13 @@ limitations under the License.
     >
       <q-icon name="warning" style="font-size: 2.5em; color: red" />
     </q-btn>
-    <v-tour
-      name="ZclTour"
-      :steps="tutorialSteps"
-      :callbacks="{ onFinish: disableTutorial, onSkip: disableTutorial }"
-    ></v-tour>
-    <q-dialog
-      v-model="deletingTutorialEndpoint"
-      class="background-color:transparent"
-    >
-      <q-card>
-        <q-card-section>
-          <div class="text-h6">Delete Endpoint</div>
-
-          Do you want to delete the endpoint used for the tutorial?
-        </q-card-section>
-        <q-card-actions>
-          <q-btn label="Cancel" v-close-popup class="col" />
-          <q-btn
-            :label="'Delete'"
-            color="primary"
-            class="col"
-            v-close-popup="deleteEndpointDialog"
-            @click="deleteEndpoint()"
-            id="delete_last_endpoint"
-          />
-        </q-card-actions>
-      </q-card>
-    </q-dialog>
+    <VueTour />
   </div>
 </template>
 
 <script>
 import { QSpinnerGears } from 'quasar'
-import CommonMixin from './util/common-mixin'
-import tutorialConfig from './tutorials/tutorialConfig.json'
-import * as Storage from './util/storage'
+import VueTour from './tutorials/VueTour.vue'
 const rendApi = require(`../src-shared/rend-api.js`)
 const observable = require('./util/observable.js')
 const dbEnum = require(`../src-shared/db-enum.js`)
@@ -105,32 +76,12 @@ function initLoad(store) {
 
 export default {
   name: 'App',
+  components: {
+    VueTour,
+  },
   computed: {
     showExceptionIcon() {
       return this.$store.state.zap.showExceptionIcon
-    },
-    endpoints: {
-      get() {
-        return Array.from(this.endpointIdListSorted.keys()).map((id) => ({
-          id: id,
-        }))
-      },
-    },
-    tourCluster: {
-      get() {
-        return this.$store.state.zap.clusterDataForTutorial
-      },
-    },
-    zclDeviceTypeOptions: {
-      get() {
-        let dt = this.$store.state.zap.zclDeviceTypes
-        let keys = Object.keys(dt).sort((a, b) => {
-          return dt[a].description.localeCompare(dt[b].description)
-        })
-        return keys.map((item) => {
-          return { deviceTypeRef: item, deviceIdentifier: dt[item].code }
-        })
-      },
     },
   },
   methods: {
@@ -155,238 +106,8 @@ export default {
       this.$store.dispatch('zap/setDefaultUiMode', 'general')
       this.$store.commit('zap/toggleShowExceptionIcon', false)
     },
-    createNewEndpointForTour(resolve) {
-      let deviceType = this.zclDeviceTypeOptions.find(
-        (item) => item.deviceIdentifier == 515
-      ).deviceTypeRef
-      if (this.endpoints.length < 1) {
-        this.$store
-          .dispatch(`zap/addEndpointType`, {
-            name: 'Anonymous Endpoint Type',
-            deviceTypeRef: deviceType,
-          })
-          .then((response) => {
-            this.$store
-              .dispatch(`zap/addEndpoint`, {
-                endpointId: 1,
-                networkId: 0,
-                profileId: parseInt('0x0107'),
-                endpointType: response.id,
-                endpointVersion: 1,
-                deviceIdentifier: 515,
-              })
-              .then((res) => {
-                if (this.shareClusterStatesAcrossEndpoints()) {
-                  this.$store.dispatch(
-                    'zap/shareClusterStatesAcrossEndpoints',
-                    {
-                      endpointTypeIdList: this.endpointTypeIdList,
-                    }
-                  )
-                }
-
-                this.$store.dispatch('zap/updateSelectedEndpointType', {
-                  endpointType: this.endpointType[res.id],
-                  deviceTypeRef:
-                    this.endpointDeviceTypeRef[this.endpointType[res.id]],
-                })
-
-                // collect all cluster id from new endpoint
-                this.selectionClients.forEach((id) => {
-                  this.updateSelectedComponentRequest({
-                    clusterId: id,
-                    side: ['client'],
-                    added: true,
-                  })
-                })
-
-                this.selectionServers.forEach((id) => {
-                  this.updateSelectedComponentRequest({
-                    clusterId: id,
-                    side: ['server'],
-                    added: true,
-                  })
-                })
-
-                this.$store.dispatch('zap/updateSelectedEndpoint', res.id)
-                resolve()
-              })
-          })
-      } else {
-        this.$store.dispatch('zap/updateSelectedEndpointType', {
-          endpointType: this.endpointType[this.endpoints[0].id],
-          deviceTypeRef:
-            this.endpointDeviceTypeRef[this.endpointType[this.endpoints[0].id]],
-        })
-        this.$store.dispatch('zap/updateSelectedEndpoint', this.endpoints[0].id)
-        resolve()
-      }
-    },
-    handleDeletionDialog() {
-      this.deletingTutorialEndpoint = true
-      this.deleteEndpointDialog = !this.deleteEndpointDialog
-    },
-
-    //  ----------  Vue Tour Functions ----------  //
-    deleteEndpoint() {
-      this.$store
-        .dispatch('zap/deleteEndpoint', this.endpoints[0].id)
-        .then(() => {
-          this.$store.dispatch(
-            'zap/deleteEndpointType',
-            this.endpointType[this.endpoints[0].id]
-          )
-        })
-      this.deletingTutorialEndpoint = false
-      this.deleteEndpointDialog = false
-      this.$store.commit('zap/openZclExtensionsDialogForTutorial', false)
-    },
-    disableTutorial() {
-      if (this.$router.currentRoute.path === '/') {
-        this.$store.commit('zap/toggleTutorial', false)
-        this.$store.commit('zap/triggerExpanded', false)
-        this.endpoints.length > 0 ? this.handleDeletionDialog() : ''
-      } else {
-        this.$router
-          .push({ name: 'Home' })
-          .then(() => {})
-          .then(() => {
-            this.endpoints.length > 0 ? this.handleDeletionDialog() : ''
-          })
-      }
-    },
-    startTutorialAndCloseTheEndpointModal() {
-      return new Promise((resolve) => {
-        this.$store.commit('zap/toggleEndpointModal', false)
-        this.$store.commit('zap/toggleTutorial', true)
-        resolve()
-      })
-    },
-    openEndpointModal() {
-      return new Promise((resolve) => {
-        this.$store.commit('zap/toggleEndpointModal', true)
-        setTimeout(() => {
-          resolve()
-        }, 250)
-      })
-    },
-    createMockEndpoint() {
-      return new Promise((resolve) => {
-        this.$store.commit('zap/toggleEndpointModal', true)
-        setTimeout(() => {
-          resolve()
-        }, 300)
-      })
-    },
-    generateEndpointCard() {
-      return new Promise((resolve) => {
-        this.createNewEndpointForTour(resolve)
-        this.$store.commit('zap/toggleEndpointModal', false)
-      })
-    },
-    expendCluster() {
-      return new Promise((resolve) => {
-        this.$store.commit('zap/triggerExpanded', true)
-        resolve()
-      })
-    },
-    comeBackToHomePage() {
-      return new Promise((resolve) => {
-        if (this.$router.currentRoute.path !== '/') {
-          this.$router.push({ name: 'Home' }).then(() => {
-            this.$store.commit('zap/triggerExpanded', true)
-            resolve()
-          })
-        } else {
-          resolve()
-        }
-      })
-    },
-    openConfigure() {
-      return new Promise((resolve) => {
-        if (this.$router.currentRoute.path === '/') {
-          console.log(this.tourCluster)
-          this.$store.commit('zap/triggerExpanded', false)
-          this.$store
-            .dispatch('zap/updateSelectedCluster', this.tourCluster)
-            .then(() => {
-              this.$store.dispatch(
-                'zap/refreshEndpointTypeCluster',
-                this.selectedEndpointTypeId
-              )
-              this.$store.dispatch('zap/setLastSelectedDomain', 'General')
-            })
-          this.$router.push({ name: 'cluster' }).then(() => {
-            resolve()
-          })
-        } else {
-          resolve()
-        }
-      })
-    },
-    backToAttributesTab() {
-      return new Promise((resolve) => {
-        this.$store.commit('zap/openReportTabInCluster', 'attributes')
-        resolve()
-      })
-    },
-    openReportTabInCluster() {
-      return new Promise((resolve) => {
-        this.$store.commit('zap/openReportTabInCluster', 'reporting')
-        resolve()
-      })
-    },
-    openCommandsTabInCluster() {
-      return new Promise((resolve) => {
-        if (this.$router.currentRoute.path === '/') {
-          this.$router.push({ name: 'cluster' }).then(() => {
-            setTimeout(() => {
-              this.$store.commit('zap/openReportTabInCluster', 'commands')
-              resolve()
-            }, 200)
-          })
-        } else {
-          this.$store.commit('zap/openReportTabInCluster', 'commands')
-          setTimeout(() => {
-            resolve()
-          }, 200)
-        }
-      })
-    },
-    backToHomePage() {
-      return new Promise((resolve) => {
-        if (this.$router.currentRoute.path !== '/') {
-          this.$store.commit('zap/openReportTabInCluster', 'attributes')
-          this.$router.push({ name: 'Home' }).then(() => {
-            resolve()
-          })
-        } else {
-          this.$store.commit('zap/openZclExtensionsDialogForTutorial', false)
-          resolve()
-        }
-      })
-    },
-    openZclExtensionDialog() {
-      return new Promise((resolve) => {
-        this.$store.commit('zap/openZclExtensionsDialogForTutorial', true)
-        setTimeout(() => {
-          resolve()
-        }, 300)
-      })
-    },
-  },
-  mixins: [CommonMixin],
-  data() {
-    return {
-      tutorialSteps: [],
-      deletingTutorialEndpoint: false,
-      deleteEndpointDialog: false,
-    }
   },
   mounted() {
-    if (this.$store.state.zap.showDevTools) {
-      this.$tours['ZclTour'].start()
-    }
     window[rendApi.GLOBAL_SYMBOL_EXECUTE](
       rendApi.id.setDarkTheme,
       storage.getItem(rendApi.storageKey.isDarkThemeActive)
@@ -444,25 +165,6 @@ export default {
     this.$onWebSocket(dbEnum.wsCategory.ucComponentStateReport, (resp) => {
       this.$store.dispatch('zap/updateUcComponentState', resp)
     })
-  },
-  created() {
-    let config = []
-    tutorialConfig.tutorialSteps.forEach((item) => {
-      config.push({
-        target: item.target,
-        header: {
-          title: item.title,
-        },
-        params: {
-          placement: item.placement,
-          enableScrolling: item.enableScrolling,
-          highlight: item.highlight,
-        },
-        content: item.content,
-        before: () => (item.before !== '' ? this[item.before]() : ''),
-      })
-    })
-    this.tutorialSteps = config
   },
 }
 </script>
