+++ src/store/zap/state.js
@@ -150,5 +150,9 @@ export default function () {
       zclSdkExtClusterToUcComponentMap: [],
     },
     allEndpointsData: {},
+    deviceTypeRefAndDeviceIdPair: {
+      deviceTypeRef: null,
+      deviceIdentifier: null,
+    },
   }
 }
