+++ src/tutorials/tutorialConfig.json
@@ -31,7 +31,7 @@
       "target": ".v-step-3",
       "title": "",
       "content": "To change the number of the endpoint on which you would like this device to appear, change the <strong>Endpoint</strong> setting",
-      "before": "",
+      "before": "setDeviceEndpoint",
       "placement": "left",
       "enableScrolling": true,
       "highlight": true
@@ -86,7 +86,7 @@
       "title": "Configuring a Cluster",
       "content": "Each cluster can be configured to implement Zigbee cluster attributes and commands. The Enable Command Discovery toggle in the top interface ribbon allows the list of commands supported by the device to be visible through the Zigbee General Command Discovery command interface.",
       "before": "comeBackToHomePage",
-      "placement": "bottom",
+      "placement": "right",
       "enableScrolling": false,
       "highlight": true
     },
@@ -140,7 +140,7 @@
       "title": "Adding Custom Clusters",
       "content": "In Zigbee, developers can add their own custom clusters to their applications. This functionality is supported by the Zigbee Cluster Library (ZCL). The custom ZCL must be described in the Silicon Labs ZCL XML format. For examples of custom ZCL XML, see the file sampleextensions.xml in the <GSDK install location>/app/zcl directory. This XML file can be used as a reference for your custom ZCL XML file. Once you have created your custom ZCL XML file, it can be added to your project through the ZCL EXTENSIONS… interface",
       "before": "backToHomePage",
-      "placement": "bottom",
+      "placement": "right",
       "enableScrolling": false,
       "highlight": true
     },
